Include dot pictodiff prototype code
Integrate Will's code as a PictoSource extension, and test a bit
with appending different viewtrees to the generated tree.

Next step will involve devising a set of methods to perform ad-hoc
comparisons between nodes in the diffed viewtrees, so that the dot
diff engine can be used to show the differences.

Also, thinking about making an "Extended dot pictosource", which
not only shows the rendered dot file as an SVG, but also the
individual viewtrees for each node (Will's code already provides
with that)