Fix: solve some extra bugs adding/modifying edges
Apart from some typos/copypasting bugs, all were related to how to
manage links when dealing with added nodes (e.g. include in the graph
those links belonging to an added node, this is, when the added node is
the source of the link).

Next: now that the big bunch of cases I could think of is well-treated,
I'll perform some refactorings that will considerably simplify the
dotdiff code and I'll start testing it with Ecore2Dot.