Fix svg mouse event bugs
Now the events cause a stroke color change, instead of a background
color one. It may be less visible, but it will be more compatible with
different visualizations (e.g. there are several badly ordered
rectangles in the generated svg of the ecore2dot transformation,
changing the background color wrongly hides some elements)