DotDiffEngine: Properly support links between ports
Svg events have been disabled for the time being (the dot to svg parser
of graphviz-java has some problems with the comps vis).

So, DotDiffEngine returns a diff in dot format that is transformed by
the standard picto pipeline. I'll investigate the problems to find the
best way to get the events support back.