Disambiguate left and right meaning in diff engines
The diff method of the engines used left and right to refer to the old
and new compared ViewTrees, respectively. With those names alone it was
not clear which version was what. Moreover, the used left and right
meaning was the opposite of what EGit compare editor does (left for
current/new version, and right for previous/old one).

Now, the diff methods use oldView and newView as parameters for a
clearer understanding.

Future commits will update the remaining diffEngines, and will swap left
and right views (of those engines where that is relevant) to offer a
consistent experience with EGit compare editor.