Make a best-effort use of the diffEngine optional parameter
If the indicated diffEngine exists and supports the current view format,
it is used. In any other case, the diff engine is automatically set.

Useful when working with views that use several formats.

Enhancement: Only one universal diffEngine can be set by using a parameter. Allow instead to change between diff engines with a menu of the view. Maybe an idea to have in mind if PictoDiff gets its own view with dedicated diff/compare controls (instead of using Picto's one).